+++ src/aceinna/devices/parsers/dmu_field_parser.py
@@ -1,3 +1,4 @@
+import math
 import struct
 
 IIR_50HZ_LPF = 50
@@ -60,7 +61,7 @@ def encode_lpf(value):
     return struct.pack('>h', counts)
 
 
-def decode_value(data_type, data):
+def do_decode_value(data_type, data):
     if data_type == 'uint64':
         try:
             pack_item = struct.pack('8B', *data)
@@ -181,6 +182,18 @@ def decode_value(data_type, data):
         return decode_lpf(data)
     else:
         return False
+
+
+def decode_value(data_type, data):
+    ret_value = do_decode_value(data_type, data)
+
+    if not isinstance(ret_value, float):
+        return ret_value
+
+    if math.isnan(ret_value):
+        return 0
+    else:
+        return ret_value
 
 
 def encode_value(data_type, data):