+++ src/aceinna/devices/ping/open.py
@@ -76,6 +76,13 @@ def ping(communicator, *args):
             'app_info': app_info_text
         }
 
+    if _need_check(filter_device_type, 'RTKL') and device_info_text.find('RTK330L') > -1:
+        return {
+            'device_type': 'RTKL',
+            'device_info': device_info_text,
+            'app_info': app_info_text
+        }
+
     if _need_check(filter_device_type, 'IMU') and device_info_text.find('OpenIMU') > -1 and \
             device_info_text.find('OpenRTK') == -1:
         return {