+++ src/aceinna/devices/upgrade_workers/__init__.py
@@ -1,2 +1,3 @@
 from .firmware_worker import FirmwareUpgradeWorker
+from .firmware_worker import EVENT_TYPE as FIRMWARE_EVENT_TYPE
 from .sdk_worker import SDKUpgradeWorker