+++ src/aceinna/devices/upgrade_workers/firmware_worker.py
@@ -4,19 +4,27 @@ from ...framework.utils import helper
 from ..ping.open import ping
 
 
+class EVENT_TYPE:
+    '''
+    Event type of Device Message Center
+    '''
+    FIRST_PACKET = 'first_packet'
+    BEFORE_WRITE = 'before_write'
+    AFTER_WRITE = 'after_write'
+
+
 class FirmwareUpgradeWorker(UpgradeWorkerBase):
     '''Firmware upgrade worker
     '''
 
-    def __init__(self, communicator, baudrate, file_content):
+    def __init__(self, communicator, baudrate, file_content, block_size=240):
         super(FirmwareUpgradeWorker, self).__init__()
         self._file_content = file_content
         self._communicator = communicator
         self.current = 0
         self.total = len(file_content)
-        self.max_data_len = 240
         self._baudrate = baudrate
-
+        self.max_data_len = block_size  # custom
         # self._key = None
         # self._is_stopped = False
 
@@ -38,8 +46,9 @@ class FirmwareUpgradeWorker(UpgradeWorkerBase):
         except Exception as ex:  # pylint: disable=broad-except
             return False
 
+        # custom
         if current == 0:
-            time.sleep(8)
+            self.emit(EVENT_TYPE.FIRST_PACKET)
 
         response = helper.read_untils_have_data(
             self._communicator, 'WA', 50, 50)
@@ -66,6 +75,8 @@ class FirmwareUpgradeWorker(UpgradeWorkerBase):
         # It is used to skip streaming data with size 1000 per read
         helper.read_untils_have_data(
             self._communicator, 'JI', 1000, 50)
+
+        self.emit(EVENT_TYPE.BEFORE_WRITE)
 
         while self.current < self.total:
             if self._is_stopped:
@@ -102,3 +113,5 @@ class FirmwareUpgradeWorker(UpgradeWorkerBase):
 
         if self.total > 0 and self.current >= self.total:
             self.emit('finish', self._key)
+
+        self.emit(EVENT_TYPE.AFTER_WRITE)