+++ src/aceinna/devices/widgets/__init__.py
@@ -1,0 +1,2 @@
+from .ntrip_client import NTRIPClient
+from .lan_data_logger import LanDataLogger