+++ src/aceinna/framework/context.py
@@ -10,6 +10,7 @@ class AppContext:
     '''
     _active_app = None
     _logger = None
+    _print_logger=None
 
     def __init__(self):
         pass
@@ -39,6 +40,14 @@ class AppContext:
         if not self._logger:
             self._logger = AppLogger(filename='default')
         return self._logger
+
+    def set_print_logger(self,logger):
+        self._print_logger = logger
+
+    def get_print_logger(self):
+        if not self._print_logger:
+            self._print_logger = AppLogger(filename='default')
+        return self._print_logger
 
 
 APP_CONTEXT = AppContext()